Fix a performance problem
ctidscan uses heap_beginscan() with the scankeys specified by the ctid
range, and then heap_getnext() returns tuples that pass the scankeys test.
Although this method internally reads all the blocks in the target table.

This adds the ctid range test into CTidAccessCustomScan(), so that ctidscan
can stop scanning the table as soon as it runs out of the ctid range.